Split sensor feature boundaries
