+++ custom_components/enphase_ev/sensor.py
@@ -12,7 +12,6 @@ from dataclasses import dataclass, field
 from datetime import datetime, timedelta, timezone
 import math
 import re
-import time
 from typing import Any, TypeVar, cast
 
 from homeassistant.components.sensor import (
@@ -39,11 +38,8 @@ from homeassistant.util import dt as dt_util
 from homeassistant.util.unit_conversion import DistanceConverter
 
 from .ac_battery_support import (
-    ac_battery_device_info,
     ac_battery_entities_available,
     ac_battery_last_reported_snapshot,
-    ac_battery_snapshot_last_reported,
-    ac_battery_storage_snapshot,
 )
 from .battery_schedule_editor import (
     BatteryScheduleRecord,
@@ -76,9 +72,49 @@ from .grid_profile_runtime import (
     SUPPORT_UNAVAILABLE,
     GridProfileRuntime,
 )
-from .parsing_helpers import coerce_optional_float, heatpump_status_text
+from .parsing_helpers import coerce_optional_float
 from .power_validation import EXTREME_SITE_POWER_W, ExtremePowerValidator
 from .runtime_data import EnphaseConfigEntry, get_runtime_data
+from .sensor_base import EnphaseSiteSensorEntity as _SiteBaseEntity
+from .sensor_battery import (
+    BATTERY_LED_STATUS_STATE_MAP as BATTERY_LED_STATUS_STATE_MAP,
+    EnphaseAcBatteryStorageChargeSensor,
+    EnphaseAcBatteryStorageCycleCountSensor,
+    EnphaseAcBatteryStorageLastReportedSensor,
+    EnphaseAcBatteryStorageOperatingModeSensor,
+    EnphaseAcBatteryStoragePowerSensor,
+    EnphaseAcBatteryStorageStatusSensor,
+    EnphaseBatteryStorageChargeSensor,
+    EnphaseBatteryStorageCycleCountSensor,
+    EnphaseBatteryStorageHealthSensor,
+    EnphaseBatteryStorageLastReportedSensor as EnphaseBatteryStorageLastReportedSensor,
+    EnphaseBatteryStorageStatusSensor,
+    _EnphaseAcBatteryStorageBaseSensor as _EnphaseAcBatteryStorageBaseSensor,
+    _EnphaseBatteryStorageBaseSensor as _EnphaseBatteryStorageBaseSensor,
+)
+from .sensor_heatpump import (
+    EnphaseHeatPumpConnectivityStatusSensor,
+    EnphaseHeatPumpDailyBatteryEnergySensor,
+    EnphaseHeatPumpDailyEnergySensor,
+    EnphaseHeatPumpDailyGridEnergySensor,
+    EnphaseHeatPumpDailySolarEnergySensor,
+    EnphaseHeatPumpEnergyMeterSensor,
+    EnphaseHeatPumpLastReportedSensor,
+    EnphaseHeatPumpPowerSensor,
+    EnphaseHeatPumpSgReadyGatewaySensor,
+    EnphaseHeatPumpSgReadyModeSensor,
+    EnphaseHeatPumpStatusSensor,
+    _heatpump_daily_snapshot as _heatpump_daily_snapshot,
+    _heatpump_member_device_type as _heatpump_member_device_type,
+    _heatpump_member_last_reported as _heatpump_member_last_reported,
+    _heatpump_member_status_text as _heatpump_member_status_text,
+    _heatpump_runtime_device_uid,
+    _heatpump_runtime_snapshot as _heatpump_runtime_snapshot,
+    _heatpump_sg_ready_semantics as _heatpump_sg_ready_semantics,
+    _heatpump_snapshot as _heatpump_snapshot,
+    _heatpump_type_snapshot as _heatpump_type_snapshot,
+    _heatpump_worst_status_text as _heatpump_worst_status_text,
+)
 from .runtime_helpers import (
     coerce_optional_text as _gateway_clean_text,
     inventory_type_available as _type_available,
@@ -92,6 +128,9 @@ from .serial_discovery import (
     active_inverter_serials_for_cleanup,
 )
 from .sensor_registry import EnphaseSensorRegistrySetup
+from .sensor_snapshot_helpers import (
+    parse_gateway_timestamp as _gateway_parse_timestamp,
+)
 from .serial_entity_metadata import (
     AC_BATTERY_ENTITY_UNIQUE_SUFFIXES as AC_BATTERY_ENTITY_UNIQUE_SUFFIXES,
     AC_BATTERY_RETIRED_UNIQUE_SUFFIXES as AC_BATTERY_RETIRED_UNIQUE_SUFFIXES,
@@ -107,6 +146,16 @@ from .tariff import (
 )
 from . import sensor_battery_helpers as _battery_helpers
 from .evse_runtime import evse_power_is_actively_charging
+
+_GATEWAY_STATUS_KEYS: tuple[str, ...] = ("statusText", "status")
+_GATEWAY_MODEL_KEYS: tuple[str, ...] = ("model", "channel_type", "sku_id")
+_GATEWAY_FIRMWARE_KEYS: tuple[str, ...] = ("envoy_sw_version", "sw_version")
+_GATEWAY_LAST_REPORT_KEYS: tuple[str, ...] = (
+    "last_report",
+    "last_reported",
+    "lastReportedAt",
+)
+_GATEWAY_IP_KEYS: tuple[str, ...] = ("ip", "ip_address", "ip-address")
 
 PARALLEL_UPDATES = 0
 
@@ -147,14 +196,6 @@ def _retain_grid_profile_sensors(coord: EnphaseCoordinator) -> bool:
     )
 
 
-BATTERY_LED_STATUS_STATE_MAP: dict[int, str] = {
-    12: "charging",
-    13: "discharging",
-    14: "idle",
-    15: "idle",
-    16: "idle",
-    17: "idle",
-}
 _battery_last_reported_members = _battery_helpers.battery_last_reported_members
 _battery_last_reported_snapshot = _battery_helpers.battery_last_reported_snapshot
 _battery_optional_bool = _battery_helpers.battery_optional_bool
@@ -3523,528 +3564,6 @@ class EnphaseInverterLifetimeEnergySensor(CoordinatorEntity, RestoreSensor):  # 
         )
 
 
-class _EnphaseBatteryStorageBaseSensor(CoordinatorEntity, SensorEntity):  # type: ignore[misc]
-    _attr_has_entity_name = True
-    _unrecorded_attributes = frozenset(
-        {"serial_number", "status", "sampled_at_utc", "state"}
-    )
-
-    def __init__(
-        self, coord: EnphaseCoordinator, serial: str, unique_suffix: str
-    ) -> None:
-        super().__init__(coord)
-        self._coord = coord
-        self._sn = str(serial)
-        self._attr_unique_id = (
-            f"{DOMAIN}_site_{coord.site_id}_battery_{self._sn}{unique_suffix}"
-        )
-        self._snapshot_cache_token: tuple[int, int] | None = None
-        self._snapshot_cache: dict[str, object] | None = None
-
-    def _snapshot(self) -> dict[str, object] | None:
-        coordinator_data = getattr(self._coord, "data", None)
-        battery_data = getattr(self._coord, "_battery_storage_data", None)
-        cacheable = coordinator_data is not None or battery_data is not None
-        token = (id(coordinator_data), id(battery_data))
-        if cacheable and token == self._snapshot_cache_token:
-            return self._snapshot_cache
-        getter = getattr(self._coord, "battery_storage", None)
-        if not callable(getter):
-            return None
-        payload = getter(self._sn)
-        if isinstance(payload, dict):
-            if cacheable:
-                self._snapshot_cache_token = token
-                self._snapshot_cache = payload
-            return payload
-        if cacheable:
-            self._snapshot_cache_token = token
-            self._snapshot_cache = None
-        return None
-
-    @staticmethod
-    def _as_int(value: object) -> int | None:
-        if value is None:
-            return None
-        try:
-            return int(str(value).strip())
-        except Exception:  # noqa: BLE001
-            return None
-
-    @staticmethod
-    def _parse_timestamp(value: object) -> datetime | None:
-        return _battery_parse_timestamp(value)
-
-    @property
-    def available(self) -> bool:
-        if not _type_available(self._coord, "encharge"):
-            return False
-        return bool(super().available and self._snapshot() is not None)
-
-    @property
-    def _battery_label(self) -> str:
-        snapshot = self._snapshot() or {}
-        for key in ("name", "serial_number", "identity"):
-            value = snapshot.get(key)
-            if value is None:
-                continue
-            try:
-                text = str(value).strip()
-            except Exception:  # noqa: BLE001
-                continue
-            if text:
-                return text
-        return self._sn
-
-    @property
-    def device_info(self) -> Any:
-        from homeassistant.helpers.entity import DeviceInfo
-
-        info = _type_device_info(self._coord, "encharge")
-        if info is not None:
-            return info
-        return DeviceInfo(
-            identifiers={(DOMAIN, f"type:{self._coord.site_id}:encharge")},
-            manufacturer="Enphase",
-            name="IQ Battery",
-        )
-
-
-class EnphaseBatteryStorageChargeSensor(_EnphaseBatteryStorageBaseSensor):
-    """Per-battery state-of-charge sensor under the shared battery type device."""
-
-    _attr_translation_key = "battery_storage_charge"
-    _attr_device_class = SensorDeviceClass.BATTERY
-    _attr_native_unit_of_measurement = "%"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_charge_level")
-
-    @property
-    def name(self) -> str:
-        return self._battery_label
-
-    @property
-    def native_value(self) -> Any:
-        snapshot = self._snapshot() or {}
-        value = snapshot.get("current_charge_pct")
-        if value is None:
-            return None
-        try:
-            return round(float(value), 1)  # type: ignore[arg-type]
-        except Exception:  # noqa: BLE001
-            return None
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = self._snapshot() or {}
-        sampled_at = _battery_snapshot_last_reported(snapshot)
-        return {
-            "serial_number": snapshot.get("serial_number") or self._sn,
-            "status": snapshot.get("status"),
-            "sampled_at_utc": (
-                sampled_at.isoformat() if sampled_at is not None else None
-            ),
-        }
-
-
-class EnphaseBatteryStorageStatusSensor(_EnphaseBatteryStorageBaseSensor):
-    _attr_translation_key = "battery_storage_status"
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_status")
-        self._attr_translation_placeholders = {"serial": self._sn}
-
-    @staticmethod
-    def _led_status_value(value: object) -> int | None:
-        if value is None or isinstance(value, bool):
-            return None
-        if isinstance(value, int):
-            return value
-        if isinstance(value, float):
-            return int(value) if value.is_integer() else None
-        try:
-            text = str(value).strip()
-        except Exception:  # noqa: BLE001
-            return None
-        if not text:
-            return None
-        try:
-            parsed = float(text)
-        except Exception:  # noqa: BLE001
-            return None
-        if not math.isfinite(parsed) or not parsed.is_integer():
-            return None
-        return int(parsed)
-
-    @property
-    def native_value(self) -> Any:
-        snapshot = self._snapshot() or {}
-        led_status = self._led_status_value(snapshot.get("led_status"))
-        return BATTERY_LED_STATUS_STATE_MAP.get(led_status, "unknown")  # type: ignore[arg-type]
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = self._snapshot() or {}
-        attrs: dict[str, object] = {}
-        led_status = self._led_status_value(snapshot.get("led_status"))
-        if led_status is not None:
-            attrs["state"] = led_status
-        return attrs
-
-
-class EnphaseBatteryStorageHealthSensor(_EnphaseBatteryStorageBaseSensor):
-    _attr_translation_key = "battery_storage_health"
-    _attr_device_class = SensorDeviceClass.BATTERY
-    _attr_native_unit_of_measurement = "%"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_health")
-        self._attr_translation_placeholders = {"serial": self._sn}
-
-    @staticmethod
-    def _parse_health_value(value: object) -> float | None:
-        if value is None:
-            return None
-        try:
-            text = str(value).strip()
-        except Exception:  # noqa: BLE001
-            return None
-        if not text:
-            return None
-        if text.endswith("%"):
-            text = text[:-1].strip()
-        if not text:
-            return None
-        try:
-            parsed = float(text)
-        except Exception:  # noqa: BLE001
-            return None
-        if not math.isfinite(parsed):
-            return None
-        return parsed
-
-    def _health_value(self) -> float | None:
-        snapshot = self._snapshot() or {}
-        for key in (
-            "battery_soh",
-            "soh",
-            "state_of_health",
-            "stateOfHealth",
-            "battery_health",
-            "health",
-        ):
-            parsed = self._parse_health_value(snapshot.get(key))
-            if parsed is not None:
-                return parsed
-        return None
-
-    @property
-    def available(self) -> bool:
-        return bool(super().available and self._health_value() is not None)
-
-    @property
-    def native_value(self) -> Any:
-        value = self._health_value()
-        if value is None:
-            return None
-        return round(value, 1)
-
-
-class EnphaseBatteryStorageCycleCountSensor(_EnphaseBatteryStorageBaseSensor):
-    _attr_translation_key = "battery_storage_cycle_count"
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_cycle_count")
-        self._attr_translation_placeholders = {"serial": self._sn}
-
-    @property
-    def native_value(self) -> Any:
-        snapshot = self._snapshot() or {}
-        return self._as_int(snapshot.get("cycle_count"))
-
-
-class EnphaseBatteryStorageLastReportedSensor(_EnphaseBatteryStorageBaseSensor):
-    _attr_translation_key = "battery_storage_last_reported"
-    _attr_device_class = SensorDeviceClass.TIMESTAMP
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_entity_registry_enabled_default = False
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_last_reported")
-        self._attr_translation_placeholders = {"serial": self._sn}
-
-    @property
-    def available(self) -> bool:
-        return bool(super().available and self.native_value is not None)
-
-    @property
-    def native_value(self) -> Any:
-        return _battery_snapshot_last_reported(self._snapshot() or {})
-
-
-class _EnphaseAcBatteryStorageBaseSensor(CoordinatorEntity, SensorEntity):  # type: ignore[misc]
-    _attr_has_entity_name = True
-    _unrecorded_attributes = frozenset(
-        {
-            "battery_id",
-            "part_number",
-            "phase",
-            "status_text",
-            "sleep_state",
-            "sleep_control_class",
-            "sleep_control_label",
-            "operating_mode",
-            "last_reported_utc",
-        }
-    )
-
-    def __init__(
-        self, coord: EnphaseCoordinator, serial: str, unique_suffix: str
-    ) -> None:
-        super().__init__(coord)
-        self._coord = coord
-        self._sn = str(serial)
-        self._attr_unique_id = (
-            f"{DOMAIN}_site_{coord.site_id}_ac_battery_{self._sn}{unique_suffix}"
-        )
-        self._snapshot_cache_token: tuple[int, int] | None = None
-        self._snapshot_cache: dict[str, object] | None = None
-
-    def _snapshot(self) -> dict[str, object] | None:
-        coordinator_data = getattr(self._coord, "data", None)
-        battery_data = getattr(self._coord, "_ac_battery_data", None)
-        cacheable = coordinator_data is not None or battery_data is not None
-        token = (id(coordinator_data), id(battery_data))
-        if cacheable and token == self._snapshot_cache_token:
-            return self._snapshot_cache
-        snapshot = ac_battery_storage_snapshot(self._coord, self._sn)
-        if cacheable:
-            self._snapshot_cache_token = token
-            self._snapshot_cache = snapshot
-        return snapshot
-
-    @staticmethod
-    def _as_int(value: object) -> int | None:
-        if value is None:
-            return None
-        try:
-            return int(round(float(str(value).strip())))
-        except Exception:  # noqa: BLE001
-            return None
-
-    @staticmethod
-    def _as_float(value: object) -> float | None:
-        if value is None:
-            return None
-        try:
-            parsed = float(value)  # type: ignore[arg-type]
-        except Exception:
-            try:
-                parsed = float(str(value).strip())
-            except Exception:  # noqa: BLE001
-                return None
-        if not math.isfinite(parsed):
-            return None
-        return parsed
-
-    @property
-    def available(self) -> bool:
-        if not _type_available(self._coord, "ac_battery"):
-            return False
-        return bool(super().available and self._snapshot() is not None)
-
-    @property
-    def _battery_label(self) -> str:
-        snapshot = self._snapshot() or {}
-        serial_number = snapshot.get("serial_number")
-        if serial_number is not None:
-            try:
-                text = str(serial_number).strip()
-            except Exception:  # noqa: BLE001
-                text = ""
-            if text:
-                return text
-        return self._sn
-
-    @property
-    def device_info(self) -> Any:
-        return ac_battery_device_info(self._coord)
-
-
-class EnphaseAcBatteryStorageChargeSensor(_EnphaseAcBatteryStorageBaseSensor):
-    _attr_translation_key = "ac_battery_storage_charge"
-    _attr_device_class = SensorDeviceClass.BATTERY
-    _attr_native_unit_of_measurement = "%"
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_charge_level")
-
-    @property
-    def name(self) -> str:
-        return self._battery_label
-
-    @property
-    def native_value(self) -> Any:
-        snapshot = self._snapshot() or {}
-        value = self._as_float(snapshot.get("current_charge_pct"))
-        if value is None:
-            return None
-        return round(value, 1)
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = dict(self._snapshot() or {})
-        return {
-            "battery_id": snapshot.get("battery_id"),
-            "part_number": snapshot.get("part_number"),
-            "phase": snapshot.get("phase"),
-            "status_text": snapshot.get("status_text"),
-            "sleep_state": snapshot.get("sleep_state"),
-        }
-
-
-class EnphaseAcBatteryStorageStatusSensor(_EnphaseAcBatteryStorageBaseSensor):
-    _attr_translation_key = "ac_battery_storage_status"
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_status")
-        self._attr_translation_placeholders = {"serial": self._sn}
-
-    @property
-    def native_value(self) -> Any:
-        snapshot = self._snapshot() or {}
-        value = snapshot.get("status_normalized")
-        if value is None:
-            return None
-        try:
-            text = str(value).strip().lower()
-        except Exception:  # noqa: BLE001
-            return None
-        return text or None
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = self._snapshot() or {}
-        return {
-            "battery_id": snapshot.get("battery_id"),
-            "status_text": snapshot.get("status_text"),
-            "sleep_state": snapshot.get("sleep_state"),
-            "sleep_control_class": snapshot.get("sleep_control_class"),
-            "sleep_control_label": snapshot.get("sleep_control_label"),
-        }
-
-
-class EnphaseAcBatteryStoragePowerSensor(_EnphaseAcBatteryStorageBaseSensor):
-    _attr_translation_key = "ac_battery_storage_power"
-    _attr_device_class = SensorDeviceClass.POWER
-    _attr_native_unit_of_measurement = UnitOfPower.WATT
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_power")
-        self._attr_translation_placeholders = {"serial": self._sn}
-
-    @property
-    def native_value(self) -> Any:
-        snapshot = self._snapshot() or {}
-        value = self._as_float(snapshot.get("power_w"))
-        if value is None:
-            return None
-        return round(value, 3)
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = self._snapshot() or {}
-        reported = ac_battery_snapshot_last_reported(snapshot)
-        return {
-            "operating_mode": snapshot.get("operating_mode"),
-            "last_reported_utc": (
-                reported.isoformat() if reported is not None else None
-            ),
-        }
-
-
-class EnphaseAcBatteryStorageOperatingModeSensor(_EnphaseAcBatteryStorageBaseSensor):
-    _attr_translation_key = "ac_battery_storage_operating_mode"
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_operating_mode")
-        self._attr_translation_placeholders = {"serial": self._sn}
-
-    @property
-    def native_value(self) -> Any:
-        snapshot = self._snapshot() or {}
-        value = snapshot.get("operating_mode")
-        if value is None:
-            return None
-        try:
-            text = str(value).strip()
-        except Exception:  # noqa: BLE001
-            return None
-        return text or None
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = self._snapshot() or {}
-        return {
-            "status_text": snapshot.get("status_text"),
-            "sleep_state": snapshot.get("sleep_state"),
-        }
-
-
-class EnphaseAcBatteryStorageCycleCountSensor(_EnphaseAcBatteryStorageBaseSensor):
-    _attr_translation_key = "ac_battery_storage_cycle_count"
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_state_class = SensorStateClass.MEASUREMENT
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_cycle_count")
-        self._attr_translation_placeholders = {"serial": self._sn}
-
-    @property
-    def native_value(self) -> Any:
-        snapshot = self._snapshot() or {}
-        return self._as_int(snapshot.get("cycle_count"))
-
-
-class EnphaseAcBatteryStorageLastReportedSensor(_EnphaseAcBatteryStorageBaseSensor):
-    _attr_translation_key = "ac_battery_storage_last_reported"
-    _attr_device_class = SensorDeviceClass.TIMESTAMP
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_entity_registry_enabled_default = False
-
-    def __init__(self, coord: EnphaseCoordinator, serial: str) -> None:
-        super().__init__(coord, serial, "_last_reported")
-        self._attr_translation_placeholders = {"serial": self._sn}
-
-    @property
-    def available(self) -> bool:
-        return bool(super().available and self.native_value is not None)
-
-    @property
-    def native_value(self) -> Any:
-        return ac_battery_snapshot_last_reported(self._snapshot() or {})
-
-
-_GATEWAY_STATUS_KEYS: tuple[str, ...] = ("statusText", "status")
-_GATEWAY_MODEL_KEYS: tuple[str, ...] = ("model", "channel_type", "sku_id")
-_GATEWAY_FIRMWARE_KEYS: tuple[str, ...] = ("envoy_sw_version", "sw_version")
-_GATEWAY_LAST_REPORT_KEYS: tuple[str, ...] = (
-    "last_report",
-    "last_reported",
-    "lastReportedAt",
-)
-_GATEWAY_IP_KEYS: tuple[str, ...] = ("ip", "ip_address", "ip-address")
-
-
 def _gateway_optional_bool(value: object) -> bool | None:
     if value is None:
         return None
@@ -4159,49 +3678,6 @@ def _gateway_summary_ip_address(
         }:
             return ip_address
     return None
-
-
-def _gateway_parse_timestamp(value: object) -> datetime | None:
-    if value is None:
-        return None
-    if isinstance(value, datetime):
-        if value.tzinfo is None:
-            return value.replace(tzinfo=timezone.utc)
-        return value.astimezone(timezone.utc)
-
-    timestamp_seconds: float | None = None
-    if isinstance(value, (int, float)):
-        try:
-            timestamp_seconds = float(value)
-        except Exception:  # noqa: BLE001
-            timestamp_seconds = None
-    elif isinstance(value, str):
-        cleaned = value.strip()
-        if not cleaned:
-            return None
-        try:
-            timestamp_seconds = float(cleaned)
-        except Exception:
-            timestamp_seconds = None
-        if timestamp_seconds is None:
-            normalized = cleaned.replace("[UTC]", "").replace("Z", "+00:00")
-            parsed = dt_util.parse_datetime(normalized)
-            if parsed is None:
-                try:
-                    parsed = datetime.fromisoformat(normalized)
-                except Exception:  # noqa: BLE001
-                    return None
-            if parsed.tzinfo is None:
-                parsed = parsed.replace(tzinfo=timezone.utc)
-            return parsed.astimezone(timezone.utc)  # type: ignore[no-any-return]
-    if timestamp_seconds is None:
-        return None
-    if timestamp_seconds > 1_000_000_000_000:
-        timestamp_seconds /= 1000.0
-    try:
-        return datetime.fromtimestamp(timestamp_seconds, tz=timezone.utc)
-    except Exception:  # noqa: BLE001
-        return None
 
 
 def _gateway_format_counts(counts: dict[str, int]) -> str | None:
@@ -4540,369 +4016,6 @@ def _microinverter_inventory_snapshot(coord: EnphaseCoordinator) -> dict[str, ob
     return snapshot  # type: ignore[no-any-return]
 
 
-def _heatpump_member_device_type(member: dict[str, object] | None) -> str | None:
-    if not isinstance(member, dict):
-        return None
-    value = (
-        member.get("device_type")
-        if member.get("device_type") is not None
-        else member.get("device-type")
-    )
-    text = _gateway_clean_text(value)
-    if not text:
-        return None
-    return text.upper()
-
-
-def _heatpump_member_status_text(member: dict[str, object] | None) -> str | None:
-    return heatpump_status_text(member)
-
-
-def _heatpump_status_counts(members: list[dict[str, object]]) -> dict[str, int]:
-    counts: dict[str, int] = {
-        "total": len(members),
-        "normal": 0,
-        "warning": 0,
-        "error": 0,
-        "not_reporting": 0,
-        "unknown": 0,
-    }
-    for member in members:
-        status_key = EnphaseCoordinator._normalize_inverter_status(
-            _heatpump_member_status_text(member)
-        )
-        counts[status_key] = int(counts.get(status_key, 0)) + 1
-    return counts
-
-
-def _heatpump_worst_status_text(status_counts: dict[str, int]) -> str | None:
-    if int(status_counts.get("error", 0) or 0) > 0:
-        return "Error"
-    if int(status_counts.get("warning", 0) or 0) > 0:
-        return "Warning"
-    if int(status_counts.get("not_reporting", 0) or 0) > 0:
-        return "Not Reporting"
-    if int(status_counts.get("unknown", 0) or 0) > 0:
-        return "Unknown"
-    if int(status_counts.get("normal", 0) or 0) > 0:
-        return "Normal"
-    return None
-
-
-def _heatpump_member_last_reported(member: dict[str, object] | None) -> datetime | None:
-    if not isinstance(member, dict):
-        return None
-    for key in ("last_report", "last_reported", "last_reported_at", "last-report"):
-        parsed = _gateway_parse_timestamp(member.get(key))
-        if parsed is not None:
-            return parsed
-    return None
-
-
-def _heatpump_snapshot(coord: EnphaseCoordinator) -> dict[str, object]:
-    summary_getter = getattr(coord, "heatpump_inventory_summary", None)
-    if callable(summary_getter):
-        try:
-            snapshot = summary_getter()
-        except Exception:  # noqa: BLE001
-            snapshot = None
-        if isinstance(snapshot, dict):
-            return snapshot
-    bucket = coord.inventory_view.type_bucket("heatpump") or {}
-    members = bucket.get("devices")
-    safe_members = (
-        [dict(item) for item in members if isinstance(item, dict)]
-        if isinstance(members, list)
-        else []
-    )
-    status_counts_raw = bucket.get("status_counts")
-    status_counts: dict[str, int] | None = None
-    if isinstance(status_counts_raw, dict):
-        parsed_counts = {
-            "total": 0,
-            "normal": 0,
-            "warning": 0,
-            "error": 0,
-            "not_reporting": 0,
-            "unknown": 0,
-        }
-        try:
-            for key in parsed_counts:
-                parsed_counts[key] = int(status_counts_raw.get(key, 0) or 0)
-            status_counts = parsed_counts
-        except Exception:
-            status_counts = None
-    if status_counts is None:
-        status_counts = _heatpump_status_counts(safe_members)
-
-    try:
-        total_devices = int(cast(Any, bucket.get("count", len(safe_members)) or 0))
-    except Exception:
-        total_devices = len(safe_members)
-    if total_devices <= 0:
-        total_devices = len(safe_members)
-    status_counts["total"] = max(int(status_counts.get("total", 0) or 0), total_devices)
-
-    latest_reported = _gateway_parse_timestamp(
-        bucket.get("latest_reported_utc")
-        if bucket.get("latest_reported_utc") is not None
-        else bucket.get("latest_reported")
-    )
-    latest_reported_device = (
-        dict(cast(dict[str, Any], bucket.get("latest_reported_device")))
-        if isinstance(bucket.get("latest_reported_device"), dict)
-        else None
-    )
-    without_last_report_count = 0
-    if latest_reported is None:
-        for member in safe_members:
-            member_last_reported = _heatpump_member_last_reported(member)
-            if member_last_reported is None:
-                without_last_report_count += 1
-                continue
-            if latest_reported is None or member_last_reported > latest_reported:
-                latest_reported = member_last_reported
-                latest_reported_device = {
-                    "device_type": _heatpump_member_device_type(member),
-                    "name": _gateway_clean_text(member.get("name")),
-                    "device_uid": _gateway_clean_text(
-                        member.get("device_uid")
-                        if member.get("device_uid") is not None
-                        else member.get("device-uid")
-                    ),
-                    "status": _heatpump_member_status_text(member),
-                }
-    overall_status_text = _gateway_clean_text(bucket.get("overall_status_text"))
-    if not overall_status_text:
-        for member in safe_members:
-            if _heatpump_member_device_type(member) != "HEAT_PUMP":
-                continue
-            overall_status_text = _heatpump_member_status_text(member)
-            if overall_status_text:
-                break
-    if not overall_status_text:
-        overall_status_text = _heatpump_worst_status_text(status_counts)
-
-    device_type_counts: dict[str, int]
-    device_type_counts_raw = bucket.get("device_type_counts")
-    if isinstance(device_type_counts_raw, dict):
-        device_type_counts = {}
-        for key, value in device_type_counts_raw.items():
-            if key is None:
-                continue
-            try:
-                count = int(value)
-            except Exception:
-                continue
-            if count <= 0:
-                continue
-            device_type_counts[str(key)] = count
-    else:
-        device_type_counts = {}
-        for member in safe_members:
-            device_type = _heatpump_member_device_type(member) or "UNKNOWN"
-            device_type_counts[device_type] = device_type_counts.get(device_type, 0) + 1
-
-    status_summary = bucket.get("status_summary")
-    if not isinstance(status_summary, str) or not status_summary.strip():
-        status_summary = EnphaseCoordinator._format_inverter_status_summary(
-            status_counts
-        )
-
-    hems_last_success_utc = getattr(coord, "_hems_devices_last_success_utc", None)
-    if not isinstance(hems_last_success_utc, datetime):
-        hems_last_success_utc = None
-    hems_last_success_mono = getattr(coord, "_hems_devices_last_success_mono", None)
-    hems_last_success_age_s: float | None = None
-    if isinstance(hems_last_success_mono, (int, float)):
-        age = time.monotonic() - float(hems_last_success_mono)
-        if age >= 0:
-            hems_last_success_age_s = round(age, 1)
-
-    return {
-        "total_devices": total_devices,
-        "members": safe_members,
-        "status_counts": status_counts,
-        "status_summary": status_summary,
-        "device_type_counts": device_type_counts,
-        "model_summary": bucket.get("model_summary"),
-        "firmware_summary": bucket.get("firmware_summary"),
-        "latest_reported": latest_reported,
-        "latest_reported_utc": (
-            latest_reported.isoformat() if latest_reported is not None else None
-        ),
-        "latest_reported_device": latest_reported_device,
-        "without_last_report_count": without_last_report_count,
-        "overall_status_text": overall_status_text,
-        "hems_data_stale": bool(getattr(coord, "_hems_devices_using_stale", False)),
-        "hems_last_success_utc": (
-            hems_last_success_utc.isoformat()
-            if hems_last_success_utc is not None
-            else None
-        ),
-        "hems_last_success_age_s": hems_last_success_age_s,
-    }
-
-
-def _heatpump_type_snapshot(
-    coord: EnphaseCoordinator, *, device_type: str
-) -> dict[str, object]:
-    summary_getter = getattr(coord, "heatpump_type_summary", None)
-    if callable(summary_getter):
-        try:
-            snapshot = summary_getter(device_type)
-        except Exception:  # noqa: BLE001
-            snapshot = None
-        if isinstance(snapshot, dict):
-            return snapshot
-    snapshot = _heatpump_snapshot(coord)
-    members = [
-        member
-        for member in snapshot.get("members", [])  # type: ignore[attr-defined]
-        if isinstance(member, dict)
-        and _heatpump_member_device_type(member) == device_type.upper()
-    ]
-    counts = _heatpump_status_counts(members)
-    latest_reported: datetime | None = None
-    latest_device: dict[str, object] | None = None
-    for member in members:
-        member_last_reported = _heatpump_member_last_reported(member)
-        if member_last_reported is None:
-            continue
-        if latest_reported is None or member_last_reported > latest_reported:
-            latest_reported = member_last_reported
-            latest_device = {
-                "name": _gateway_clean_text(member.get("name")),
-                "device_uid": _gateway_clean_text(
-                    member.get("device_uid")
-                    if member.get("device_uid") is not None
-                    else member.get("device-uid")
-                ),
-                "status": _heatpump_member_status_text(member),
-            }
-    status_texts = [
-        status
-        for status in (_heatpump_member_status_text(member) for member in members)
-        if status
-    ]
-    unique_statuses = list(dict.fromkeys(status_texts))
-    if len(unique_statuses) == 1:
-        native_status = unique_statuses[0]
-    else:
-        native_status = _heatpump_worst_status_text(counts)  # type: ignore[assignment]
-    return {
-        "device_type": device_type.upper(),
-        "members": members,
-        "member_count": len(members),
-        "status_counts": counts,
-        "status_summary": EnphaseCoordinator._format_inverter_status_summary(counts),
-        "native_status": native_status,
-        "latest_reported": latest_reported,
-        "latest_reported_utc": (
-            latest_reported.isoformat() if latest_reported is not None else None
-        ),
-        "latest_reported_device": latest_device,
-        "hems_data_stale": snapshot.get("hems_data_stale"),
-        "hems_last_success_utc": snapshot.get("hems_last_success_utc"),
-        "hems_last_success_age_s": snapshot.get("hems_last_success_age_s"),
-    }
-
-
-def _heatpump_runtime_snapshot(coord: EnphaseCoordinator) -> dict[str, object]:
-    snapshot = getattr(coord, "heatpump_runtime_state", None)
-    if isinstance(snapshot, dict):
-        return dict(snapshot)
-    return {}
-
-
-def _heatpump_daily_snapshot(coord: EnphaseCoordinator) -> dict[str, object]:
-    snapshot = getattr(coord, "heatpump_daily_consumption", None)
-    if isinstance(snapshot, dict):
-        return dict(snapshot)
-    return {}
-
-
-def _heatpump_runtime_device_uid(coord: EnphaseCoordinator) -> str | None:
-    getter = getattr(coord, "_heatpump_runtime_device_uid", None)
-    if callable(getter):
-        try:
-            return _gateway_clean_text(getter())
-        except Exception:  # noqa: BLE001
-            return None
-    return None
-
-
-def _heatpump_runtime_last_reported(snapshot: dict[str, object]) -> datetime | None:
-    return _gateway_parse_timestamp(
-        snapshot.get("last_report_at")
-        if snapshot.get("last_report_at") is not None
-        else snapshot.get("last_reported_at")
-    )
-
-
-def _heatpump_runtime_common_attrs(
-    coord: EnphaseCoordinator, snapshot: dict[str, object]
-) -> dict[str, object]:
-    last_reported = _heatpump_runtime_last_reported(snapshot)
-    return {
-        "device_uid": snapshot.get("device_uid"),
-        "member_name": snapshot.get("member_name"),
-        "member_device_type": snapshot.get("member_device_type"),
-        "pairing_status": snapshot.get("pairing_status"),
-        "device_state": snapshot.get("device_state"),
-        "runtime_endpoint_type": snapshot.get("endpoint_type"),
-        "runtime_endpoint_timestamp": snapshot.get("endpoint_timestamp"),
-        "last_report_at_utc": (
-            last_reported.isoformat() if last_reported is not None else None
-        ),
-        "source": snapshot.get("source"),
-        "using_stale": bool(
-            getattr(coord, "heatpump_runtime_state_using_stale", False)
-        ),
-        "last_success_utc": (
-            last_success.isoformat()
-            if (last_success := coord.heatpump_runtime_state_last_success_utc)
-            is not None
-            else None
-        ),
-        "last_error": getattr(coord, "heatpump_runtime_state_last_error", None),
-    }
-
-
-def _heatpump_daily_common_attrs(
-    coord: EnphaseCoordinator, snapshot: dict[str, object]
-) -> dict[str, object]:
-    return {
-        "sampled_at_utc": snapshot.get("sampled_at_utc"),
-        "device_uid": snapshot.get("device_uid"),
-        "device_name": snapshot.get("device_name"),
-        "split_device_uid": snapshot.get("split_device_uid"),
-        "split_device_name": snapshot.get("split_device_name"),
-        "member_name": snapshot.get("member_name"),
-        "member_device_type": snapshot.get("member_device_type"),
-        "pairing_status": snapshot.get("pairing_status"),
-        "device_state": snapshot.get("device_state"),
-        "daily_endpoint_type": snapshot.get("endpoint_type"),
-        "daily_endpoint_timestamp": snapshot.get("endpoint_timestamp"),
-        "split_endpoint_type": snapshot.get("split_endpoint_type"),
-        "split_endpoint_timestamp": snapshot.get("split_endpoint_timestamp"),
-        "day_key": snapshot.get("day_key"),
-        "timezone": snapshot.get("timezone"),
-        "source": snapshot.get("source"),
-        "split_source": snapshot.get("split_source"),
-        "using_stale": bool(
-            getattr(coord, "heatpump_daily_consumption_using_stale", False)
-        ),
-        "last_success_utc": (
-            last_success.isoformat()
-            if (last_success := coord.heatpump_daily_consumption_last_success_utc)
-            is not None
-            else None
-        ),
-        "last_error": getattr(coord, "heatpump_daily_consumption_last_error", None),
-    }
-
-
 def _title_case_status(value: object, hass: object | None = None) -> str | None:
     return status_label(value, hass=hass) or friendly_status_text(value)
 
@@ -5021,26 +4134,6 @@ def _gateway_terminal_values(member: dict[str, object] | None) -> dict[str, obje
         if isinstance(raw_value, (int, float, bool)):
             values[key] = raw_value
     return values
-
-
-def _heatpump_sg_ready_semantics(status_text: object) -> dict[str, object]:
-    text = _gateway_clean_text(status_text)
-    if not text:
-        return {}
-    normalized = text.casefold()
-    if normalized in {"recommended", "mode_3", "mode3"}:
-        return {
-            "sg_ready_mode": 3,
-            "sg_ready_contact_state": "closed",
-            "status_explanation": "Recommended means the SG Ready contact is closed.",
-        }
-    if normalized in {"normal", "mode_2", "mode2"}:
-        return {
-            "sg_ready_mode": 2,
-            "sg_ready_contact_state": "open",
-            "status_explanation": "Normal means the SG Ready contacts are open.",
-        }
-    return {}
 
 
 def _gateway_iq_energy_router_inventory_buckets(
@@ -5495,105 +4588,6 @@ def _gateway_dry_contact_members(
         )
     )
     return members_out
-
-
-class _SiteBaseEntity(CoordinatorEntity, SensorEntity):  # type: ignore[misc]
-    _attr_has_entity_name = True
-    _unrecorded_attributes = frozenset(
-        {
-            "last_success_utc",
-            "last_failure_utc",
-            "backoff_ends_utc",
-            "last_failure_response",
-        }
-    )
-
-    def __init__(
-        self,
-        coord: EnphaseCoordinator,
-        key: str,
-        _name: str,
-        type_key: str | None = "envoy",
-    ):
-        super().__init__(coord)
-        self._coord = coord
-        self._key = key
-        self._type_key = type_key
-        self._attr_unique_id = f"{DOMAIN}_site_{coord.site_id}_{key}"
-
-    @property
-    def available(self) -> bool:
-        if self._type_key is not None and not _type_available(
-            self._coord, self._type_key
-        ):
-            return False
-        if self._coord.last_success_utc is not None:
-            return True
-        return super().available  # type: ignore[no-any-return]
-
-    def _cloud_diag_attrs(
-        self, *, include_last_success: bool = True
-    ) -> dict[str, object]:
-        attrs: dict[str, object] = {}
-        # These attributes are intentionally sanitized by the coordinator
-        # before they become visible on diagnostic sensors.
-        if include_last_success and self._coord.last_success_utc:
-            attrs["last_success_utc"] = self._coord.last_success_utc.isoformat()
-        if self._coord.last_failure_utc:
-            attrs["last_failure_utc"] = self._coord.last_failure_utc.isoformat()
-        if self._coord.last_failure_status is not None:
-            attrs["last_failure_status"] = self._coord.last_failure_status
-        if self._coord.last_failure_description:
-            attrs["code_description"] = self._coord.last_failure_description
-        if self._coord.last_failure_response:
-            attrs["last_failure_response"] = self._coord.last_failure_response
-        if self._coord.last_failure_source:
-            attrs["last_failure_source"] = self._coord.last_failure_source
-        last_failure_endpoint = getattr(self._coord, "last_failure_endpoint", None)
-        if last_failure_endpoint:
-            attrs["last_failure_endpoint"] = last_failure_endpoint
-        payload_failure_kind = getattr(self._coord, "payload_failure_kind", None)
-        if payload_failure_kind:
-            attrs["payload_failure_kind"] = payload_failure_kind
-        payload_using_stale = bool(getattr(self._coord, "payload_using_stale", False))
-        if payload_using_stale:
-            attrs["payload_using_stale"] = True
-        if self._coord.backoff_ends_utc:
-            attrs["backoff_ends_utc"] = self._coord.backoff_ends_utc.isoformat()
-        return attrs
-
-    def _backoff_remaining_seconds(self) -> int | None:
-        ends = self._coord.backoff_ends_utc
-        if ends is None:
-            return None
-        try:
-            remaining = (ends - dt_util.utcnow()).total_seconds()
-        except Exception:
-            return None
-        if remaining <= 0:
-            return 0
-        rounded = int(round(remaining))
-        if rounded <= 0:
-            return 1
-        return rounded
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        return self._cloud_diag_attrs()
-
-    @property
-    def device_info(self) -> Any:
-        if self._type_key is None:
-            return _cloud_device_info(self._coord.site_id)
-        info = _type_device_info(self._coord, self._type_key)
-        if info is not None:
-            return info
-        from homeassistant.helpers.entity import DeviceInfo
-
-        return DeviceInfo(
-            identifiers={(DOMAIN, f"type:{self._coord.site_id}:{self._type_key}")},
-            manufacturer="Enphase",
-        )
 
 
 class EnphaseSiteEnergySensor(_SiteBaseEntity, RestoreSensor):  # type: ignore[misc]
@@ -8053,509 +7047,6 @@ class EnphaseMicroinverterLastReportedSensor(_SiteBaseEntity):
         return {
             "latest_reported_device": snapshot.get("latest_reported_device"),
         }
-
-
-class EnphaseHeatPumpStatusSensor(_SiteBaseEntity):
-    _attr_translation_key = "heat_pump_status"
-    _attr_entity_registry_enabled_default = True
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(
-            coord, "heat_pump_status", "Heat Pump Status", type_key="heatpump"
-        )
-
-    def _snapshot(self) -> dict[str, object]:
-        return _heatpump_runtime_snapshot(self._coord)
-
-    def _runtime_device_uid(self) -> str | None:
-        getter = getattr(self._coord, "_heatpump_runtime_device_uid", None)
-        if callable(getter):
-            try:
-                return getter()  # type: ignore[no-any-return]
-            except Exception:  # noqa: BLE001
-                return None
-        uid = self._snapshot().get("device_uid")
-        return _gateway_clean_text(uid)
-
-    @property
-    def available(self) -> bool:
-        if not super().available:
-            return False
-        if not self._runtime_device_uid():
-            return False
-        return self.native_value is not None
-
-    @property
-    def native_value(self) -> Any:
-        return _title_case_status(
-            self._snapshot().get("heatpump_status"),
-            getattr(self, "hass", None) or self._coord.hass,
-        )
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = self._snapshot()
-        sg_ready_details = _heatpump_sg_ready_semantics(
-            snapshot.get("sg_ready_mode_label") or snapshot.get("sg_ready_mode_raw")
-        )
-        attrs = _heatpump_runtime_common_attrs(self._coord, snapshot)
-        attrs.update(
-            {
-                "heatpump_status_raw": snapshot.get("heatpump_status"),
-                "sg_ready_mode_raw": snapshot.get("sg_ready_mode_raw"),
-                "sg_ready_mode_label": snapshot.get("sg_ready_mode_label"),
-                "sg_ready_active": snapshot.get("sg_ready_active"),
-                "sg_ready_contact_state": snapshot.get("sg_ready_contact_state"),
-                "vpp_sgready_mode_override": snapshot.get("vpp_sgready_mode_override"),
-                **sg_ready_details,
-            }
-        )
-        return attrs
-
-
-class EnphaseHeatPumpConnectivityStatusSensor(_SiteBaseEntity):
-    _attr_translation_key = "heat_pump_connectivity_status"
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_entity_registry_enabled_default = True
-    _unrecorded_attributes = _SiteBaseEntity._unrecorded_attributes.union(
-        {"members", "latest_reported_device"}
-    )
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(
-            coord,
-            "heat_pump_connectivity_status",
-            "Heat Pump Connectivity Status",
-            type_key="heatpump",
-        )
-
-    @property
-    def available(self) -> bool:
-        if not super().available:
-            return False
-        snapshot = _heatpump_snapshot(self._coord)
-        if int(snapshot.get("total_devices", 0) or 0) > 0:  # type: ignore[call-overload]
-            return True
-        return not bool(getattr(self._coord, "_devices_inventory_ready", False))
-
-    @property
-    def native_value(self) -> Any:
-        return _title_case_status(
-            _heatpump_snapshot(self._coord).get("overall_status_text"),
-            getattr(self, "hass", None) or self._coord.hass,
-        )
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = _heatpump_snapshot(self._coord)
-        members = snapshot.get("members")
-        safe_members = (
-            [dict(member) for member in members if isinstance(member, dict)]
-            if isinstance(members, list)
-            else []
-        )
-        return {
-            "total_devices": snapshot.get("total_devices"),
-            "status_counts": snapshot.get("status_counts"),
-            "status_summary": snapshot.get("status_summary"),
-            "device_type_counts": snapshot.get("device_type_counts"),
-            "model_summary": snapshot.get("model_summary"),
-            "firmware_summary": snapshot.get("firmware_summary"),
-            "latest_reported_utc": snapshot.get("latest_reported_utc"),
-            "latest_reported_device": snapshot.get("latest_reported_device"),
-            "hems_data_stale": snapshot.get("hems_data_stale"),
-            "hems_last_success_utc": snapshot.get("hems_last_success_utc"),
-            "hems_last_success_age_s": snapshot.get("hems_last_success_age_s"),
-            "members": safe_members,
-        }
-
-
-class _EnphaseHeatPumpDeviceTypeSensor(_SiteBaseEntity):
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_entity_registry_enabled_default = True
-    _unrecorded_attributes = _SiteBaseEntity._unrecorded_attributes.union(
-        {"members", "latest_reported_device"}
-    )
-    _heatpump_device_type: str
-
-    def __init__(
-        self,
-        coord: EnphaseCoordinator,
-        *,
-        key: str,
-        name: str,
-        device_type: str,
-    ) -> None:
-        super().__init__(coord, key, name, type_key="heatpump")
-        self._heatpump_device_type = device_type
-
-    def _snapshot(self) -> dict[str, object]:
-        return _heatpump_type_snapshot(
-            self._coord, device_type=self._heatpump_device_type
-        )
-
-    @property
-    def available(self) -> bool:
-        if not super().available:
-            return False
-        return int(self._snapshot().get("member_count", 0) or 0) > 0  # type: ignore[call-overload,no-any-return]
-
-    @property
-    def native_value(self) -> Any:
-        return self._snapshot().get("native_status")
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = self._snapshot()
-        members = snapshot.get("members")
-        return {
-            "device_type": snapshot.get("device_type"),
-            "member_count": snapshot.get("member_count"),
-            "status_counts": snapshot.get("status_counts"),
-            "status_summary": snapshot.get("status_summary"),
-            "latest_reported_utc": snapshot.get("latest_reported_utc"),
-            "latest_reported_device": snapshot.get("latest_reported_device"),
-            "hems_data_stale": snapshot.get("hems_data_stale"),
-            "hems_last_success_utc": snapshot.get("hems_last_success_utc"),
-            "hems_last_success_age_s": snapshot.get("hems_last_success_age_s"),
-            "members": (
-                [dict(member) for member in members if isinstance(member, dict)]
-                if isinstance(members, list)
-                else []
-            ),
-        }
-
-
-class EnphaseHeatPumpSgReadyModeSensor(_SiteBaseEntity):
-    _attr_translation_key = "heat_pump_sg_ready_mode"
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_entity_registry_enabled_default = True
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(
-            coord,
-            "heat_pump_sg_ready_mode",
-            "Heat Pump SG-Ready Mode",
-            type_key="heatpump",
-        )
-
-    def _snapshot(self) -> dict[str, object]:
-        return _heatpump_runtime_snapshot(self._coord)
-
-    def _runtime_device_uid(self) -> str | None:
-        getter = getattr(self._coord, "_heatpump_runtime_device_uid", None)
-        if callable(getter):
-            try:
-                return getter()  # type: ignore[no-any-return]
-            except Exception:  # noqa: BLE001
-                return None
-        uid = self._snapshot().get("device_uid")
-        return _gateway_clean_text(uid)
-
-    @property
-    def available(self) -> bool:
-        if not super().available:
-            return False
-        if not self._runtime_device_uid():
-            return False
-        snapshot = self._snapshot()
-        return any(
-            snapshot.get(key) is not None
-            for key in ("sg_ready_mode_label", "sg_ready_mode_raw")
-        )
-
-    @property
-    def native_value(self) -> Any:
-        snapshot = self._snapshot()
-        return snapshot.get("sg_ready_mode_label") or snapshot.get("sg_ready_mode_raw")
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = self._snapshot()
-        details = _heatpump_sg_ready_semantics(
-            snapshot.get("sg_ready_mode_label") or snapshot.get("sg_ready_mode_raw")
-        )
-        return {
-            "heatpump_status_raw": snapshot.get("heatpump_status"),
-            "sg_ready_mode_raw": snapshot.get("sg_ready_mode_raw"),
-            "sg_ready_mode_label": snapshot.get("sg_ready_mode_label"),
-            "sg_ready_active": snapshot.get("sg_ready_active"),
-            "sg_ready_contact_state": snapshot.get("sg_ready_contact_state"),
-            "vpp_sgready_mode_override": snapshot.get("vpp_sgready_mode_override"),
-            **details,
-        }
-
-
-class EnphaseHeatPumpEnergyMeterSensor(_EnphaseHeatPumpDeviceTypeSensor):
-    _attr_translation_key = "heat_pump_energy_meter"
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(
-            coord,
-            key="heat_pump_energy_meter",
-            name="Heat Pump Energy Meter Status",
-            device_type="ENERGY_METER",
-        )
-
-
-class EnphaseHeatPumpSgReadyGatewaySensor(_EnphaseHeatPumpDeviceTypeSensor):
-    _attr_translation_key = "heat_pump_sg_ready_gateway"
-    _attr_entity_registry_enabled_default = False
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(
-            coord,
-            key="heat_pump_sg_ready_gateway",
-            name="Heat Pump SG-Ready Gateway Status",
-            device_type="SG_READY_GATEWAY",
-        )
-
-
-class EnphaseHeatPumpLastReportedSensor(_SiteBaseEntity):
-    _attr_translation_key = "heat_pump_last_reported"
-    _attr_device_class = SensorDeviceClass.TIMESTAMP
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_entity_registry_enabled_default = False
-    _unrecorded_attributes = _SiteBaseEntity._unrecorded_attributes.union(
-        {"latest_reported_device"}
-    )
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(
-            coord,
-            "heat_pump_last_reported",
-            "Heat Pump Last Reported",
-            type_key="heatpump",
-        )
-
-    @property
-    def available(self) -> bool:
-        if not super().available:
-            return False
-        if _heatpump_runtime_device_uid(self._coord) is None:
-            return False
-        return (
-            _heatpump_runtime_last_reported(_heatpump_runtime_snapshot(self._coord))
-            is not None
-        )
-
-    @property
-    def native_value(self) -> Any:
-        return _heatpump_runtime_last_reported(_heatpump_runtime_snapshot(self._coord))
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        return {}
-
-
-class _EnphaseHeatPumpDailyEnergySensor(_SiteBaseEntity):
-    _attr_device_class = SensorDeviceClass.ENERGY
-    _attr_native_unit_of_measurement = UnitOfEnergy.KILO_WATT_HOUR
-    _attr_state_class = SensorStateClass.TOTAL
-    _attr_suggested_display_precision = 3
-    _daily_key: str
-
-    def __init__(self, coord: EnphaseCoordinator, key: str, name: str) -> None:
-        super().__init__(coord, key, name, type_key="heatpump")
-
-    def _snapshot(self) -> dict[str, object]:
-        return _heatpump_daily_snapshot(self._coord)
-
-    @property
-    def available(self) -> bool:
-        if not super().available:
-            return False
-        snapshot = self._snapshot()
-        return snapshot.get(self._daily_key) is not None
-
-    @property
-    def native_value(self) -> Any:
-        snapshot = self._snapshot()
-        value = snapshot.get(self._daily_key)
-        if value is None:
-            return None
-        try:
-            return round(float(value) / 1000.0, 3)  # type: ignore[arg-type]
-        except Exception:  # noqa: BLE001
-            return None
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        snapshot = self._snapshot()
-        attrs = _heatpump_daily_common_attrs(self._coord, snapshot)
-        if self._daily_key == "daily_energy_wh":
-            attrs["source"] = snapshot.get("source")
-            attrs["device_uid"] = snapshot.get("device_uid")
-            attrs["device_name"] = snapshot.get("device_name")
-        else:
-            attrs["source"] = snapshot.get("split_source") or snapshot.get("source")
-            attrs["device_uid"] = snapshot.get("split_device_uid")
-            attrs["device_name"] = snapshot.get("split_device_name")
-            attrs["using_stale"] = bool(
-                getattr(self._coord, "heatpump_daily_split_using_stale", False)
-            )
-            attrs["last_success_utc"] = (
-                last_success.isoformat()
-                if (last_success := self._coord.heatpump_daily_split_last_success_utc)
-                is not None
-                else None
-            )
-            attrs["last_error"] = getattr(
-                self._coord, "heatpump_daily_split_last_error", None
-            )
-        attrs["details"] = (
-            list(snapshot.get("details"))  # type: ignore[call-overload]
-            if isinstance(snapshot.get("details"), list)
-            else []
-        )
-        attrs["daily_energy_wh"] = snapshot.get("daily_energy_wh")
-        attrs["split_daily_energy_wh"] = snapshot.get("split_daily_energy_wh")
-        attrs["daily_grid_wh"] = snapshot.get("daily_grid_wh")
-        attrs["daily_solar_wh"] = snapshot.get("daily_solar_wh")
-        attrs["daily_battery_wh"] = snapshot.get("daily_battery_wh")
-        return attrs
-
-
-class EnphaseHeatPumpDailyEnergySensor(_EnphaseHeatPumpDailyEnergySensor):
-    _attr_translation_key = "heat_pump_daily_energy"
-    _daily_key = "daily_energy_wh"
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(coord, "heat_pump_daily_energy", "Heat Pump Daily Energy")
-
-
-class EnphaseHeatPumpDailyGridEnergySensor(_EnphaseHeatPumpDailyEnergySensor):
-    _attr_translation_key = "heat_pump_daily_grid_energy"
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_entity_registry_enabled_default = False
-    _daily_key = "daily_grid_wh"
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(
-            coord,
-            "heat_pump_daily_grid_energy",
-            "Heat Pump Daily Grid Energy",
-        )
-
-
-class EnphaseHeatPumpDailySolarEnergySensor(_EnphaseHeatPumpDailyEnergySensor):
-    _attr_translation_key = "heat_pump_daily_solar_energy"
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_entity_registry_enabled_default = False
-    _daily_key = "daily_solar_wh"
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(
-            coord,
-            "heat_pump_daily_solar_energy",
-            "Heat Pump Daily Solar Energy",
-        )
-
-
-class EnphaseHeatPumpDailyBatteryEnergySensor(_EnphaseHeatPumpDailyEnergySensor):
-    _attr_translation_key = "heat_pump_daily_battery_energy"
-    _attr_entity_category = EntityCategory.DIAGNOSTIC
-    _attr_entity_registry_enabled_default = False
-    _daily_key = "daily_battery_wh"
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(
-            coord,
-            "heat_pump_daily_battery_energy",
-            "Heat Pump Daily Battery Energy",
-        )
-
-
-class EnphaseHeatPumpPowerSensor(_SiteBaseEntity):
-    _attr_translation_key = "heat_pump_power"
-    _attr_device_class = SensorDeviceClass.POWER
-    _attr_native_unit_of_measurement = UnitOfPower.WATT
-    _attr_state_class = SensorStateClass.MEASUREMENT
-    _attr_entity_registry_enabled_default = True
-
-    def __init__(self, coord: EnphaseCoordinator) -> None:
-        super().__init__(
-            coord, "heat_pump_power", "Heat Pump Power", type_key="heatpump"
-        )
-
-    @property
-    def available(self) -> bool:
-        if not super().available:
-            return False
-        return self._coord.heatpump_power_w is not None
-
-    @property
-    def native_value(self) -> Any:
-        value = self._coord.heatpump_power_w
-        if value is None:
-            return None
-        return round(value, 3)
-
-    @property
-    def extra_state_attributes(self) -> Any:
-        runtime_snapshot = _heatpump_runtime_snapshot(self._coord)
-        daily_snapshot = _heatpump_daily_snapshot(self._coord)
-        attrs: dict[str, object] = {
-            "sampled_at_utc": (
-                self._coord.heatpump_power_sample_utc.isoformat()
-                if self._coord.heatpump_power_sample_utc is not None
-                else None
-            ),
-            "series_start_utc": (
-                self._coord.heatpump_power_start_utc.isoformat()
-                if self._coord.heatpump_power_start_utc is not None
-                else None
-            ),
-            "device_uid": self._coord.heatpump_power_device_uid,
-            "source": self._coord.heatpump_power_source,
-            "raw_power_w": self._coord.heatpump_power_raw_w,
-            "power_window_seconds": self._coord.heatpump_power_window_seconds,
-            "power_validation": self._coord.heatpump_power_validation,
-            "smoothed": self._coord.heatpump_power_smoothed,
-            "using_stale": bool(
-                getattr(self._coord, "heatpump_power_using_stale", False)
-            ),
-            "last_success_utc": (
-                last_success.isoformat()
-                if (last_success := self._coord.heatpump_power_last_success_utc)
-                is not None
-                else None
-            ),
-            "last_error": self._coord.heatpump_power_last_error,
-        }
-        attrs.update(
-            {
-                key: value
-                for key, value in _heatpump_runtime_common_attrs(
-                    self._coord, runtime_snapshot
-                ).items()
-                if key
-                not in {"device_uid", "source", "last_error", "last_report_at_utc"}
-            }
-        )
-        attrs.update(
-            {
-                key: value
-                for key, value in _heatpump_daily_common_attrs(
-                    self._coord, daily_snapshot
-                ).items()
-                if key
-                in {
-                    "device_name",
-                    "member_name",
-                    "member_device_type",
-                    "pairing_status",
-                    "device_state",
-                    "daily_endpoint_type",
-                    "daily_endpoint_timestamp",
-                    "day_key",
-                    "timezone",
-                }
-            }
-        )
-        if self._coord.heatpump_power_last_error:
-            attrs["last_error"] = self._coord.heatpump_power_last_error
-        return attrs
 
 
 class EnphaseStormAlertSensor(_SiteBaseEntity):
